Refactor - separate Connection from Ngrok
